pyproject.toml
[project]
name = "fastrtc-groq"
version = "0.1.0"
description = "RenataAI voice support assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastrtc[stt,tts,vad]>=0.0.19",
    "groq>=0.22.0",
    "numpy>=2.1.3",
    "langgraph>=0.1.18",
    "langchain-core>=0.1.29",
    "langchain-groq>=0.1.5",
    "loguru>=0.7.3",
    "dotenv>=0.9.9",
    "langchain>=0.3.27",
    "kokoro-onnx>=0.4.9",
    "twilio>=9.8.7",
    "sentence-transformers>=5.1.2",
    "faiss-cpu>=1.13.0",
]
